+++ src/Capability/State/Internal/Strategies.hs
@@ -30,7 +30,9 @@ module Capability.State.Internal.Strategies
 import Capability.Accessors
 import Capability.Reader.Internal.Class
 import Capability.State.Internal.Class
-import Control.Lens (set, view)
+import Capability.State.Internal.Strategies.Common
+import Capability.Source.Internal.Strategies ()
+import Capability.Sink.Internal.Strategies ()
 import Control.Monad.IO.Class (MonadIO, liftIO)
 import Control.Monad.Primitive (PrimMonad)
 import qualified Control.Monad.State.Class as State
@@ -42,16 +44,7 @@ import Data.IORef
 import Data.Mutable
 import GHC.Exts (Proxy#)
 
--- | Derive 'HasState' from @m@'s
--- 'Control.Monad.State.Class.MonadState' instance.
-newtype MonadState (m :: * -> *) (a :: *) = MonadState (m a)
-  deriving (Functor, Applicative, Monad, MonadIO, PrimMonad)
-
 instance State.MonadState s m => HasState tag s (MonadState m) where
-  get_ _ = coerce @(m s) State.get
-  {-# INLINE get_ #-}
-  put_ _ = coerce @(s -> m ()) State.put
-  {-# INLINE put_ #-}
   state_ :: forall a. Proxy# tag -> (s -> (a, s)) -> MonadState m a
   state_ _ = coerce @((s -> (a, s)) -> m a) State.state
   {-# INLINE state_ #-}
@@ -62,20 +55,12 @@ instance
   , forall x y. Coercible x y => Coercible (m x) (m y) )
   => HasState tag to (Coerce to m)
   where
-    get_ tag = coerce @(m from) $ get_ tag
-    {-# INLINE get_ #-}
-    put_ tag = coerce @(from -> m ()) $ put_ tag
-    {-# INLINE put_ #-}
     state_ :: forall a. Proxy# tag -> (to -> (a, to)) -> Coerce to m a
     state_ tag = coerce @((from -> (a, from)) -> m a) $ state_ tag
     {-# INLINE state_ #-}
 
 -- | Rename the tag.
 instance HasState oldtag s m => HasState newtag s (Rename oldtag m) where
-  get_ _ = coerce @(m s) $ get @oldtag
-  {-# INLINE get_ #-}
-  put_ _ = coerce @(s -> m ()) $ put @oldtag
-  {-# INLINE put_ #-}
   state_ :: forall a. Proxy# newtag -> (s -> (a, s)) -> Rename oldtag m a
   state_ _ = coerce @((s -> (a, s)) -> m a) $ state @oldtag
   {-# INLINE state_ #-}
@@ -88,12 +73,6 @@ instance
   ( tag ~ field, Generic.HasField' field record v, HasState oldtag record m )
   => HasState tag v (Field field oldtag m)
   where
-    get_ _ = coerce @(m v) $
-      gets @oldtag $ view (Generic.field' @field)
-    {-# INLINE get_ #-}
-    put_ _ = coerce @(v -> m ()) $
-      modify @oldtag . set (Generic.field' @field @record)
-    {-# INLINE put_ #-}
     state_ :: forall a.
       Proxy# tag
       -> (v -> (a, v))
@@ -110,12 +89,6 @@ instance
   ( tag ~ pos, Generic.HasPosition' pos struct v, HasState oldtag struct m )
   => HasState tag v (Pos pos oldtag m)
   where
-    get_ _ = coerce @(m v) $
-      gets @oldtag $ view (Generic.position' @pos)
-    {-# INLINE get_ #-}
-    put_ _ = coerce @(v -> m ()) $
-      modify @oldtag . set (Generic.position' @pos @struct)
-    {-# INLINE put_ #-}
     state_ :: forall a.
       Proxy# tag
       -> (v -> (a, v))
@@ -128,10 +101,6 @@ instance
 instance (HasState tag s m, MonadTrans t, Monad (t m))
   => HasState tag s (Lift (t m))
   where
-    get_ _ = coerce $ lift @t @m $ get @tag @s
-    {-# INLINE get_ #-}
-    put_ _ = coerce $ lift @t @m . put @tag @s
-    {-# INLINE put_ #-}
     state_ :: forall a. Proxy# tag -> (s -> (a, s)) -> Lift (t m) a
     state_ _ = coerce $ lift @t @m . state @tag @s @m @a
     {-# INLINE state_ #-}
@@ -143,31 +112,10 @@ deriving via ((t2 :: (* -> *) -> * -> *) ((t1 :: (* -> *) -> * -> *) m))
   , Monad m, HasState tag s (t2 (t1 m)) )
   => HasState tag s ((t2 :.: t1) m)
 
--- | Derive a state monad from a reader over an 'Data.IORef.IORef'.
---
--- Example:
---
--- > newtype MyState m a = MyState (ReaderT (IORef Int) m a)
--- >   deriving (Functor, Applicative, Monad)
--- >   deriving HasState "foo" Int via
--- >     ReaderIORef (MonadReader (ReaderT (IORef Int) m))
---
--- See 'ReaderRef' for a more generic strategy.
-newtype ReaderIORef m a = ReaderIORef (m a)
-  deriving (Functor, Applicative, Monad)
-
 instance
   (HasReader tag (IORef s) m, MonadIO m)
   => HasState tag s (ReaderIORef m)
   where
-    get_ _ = ReaderIORef $ do
-      ref <- ask @tag
-      liftIO $ readIORef ref
-    {-# INLINE get_ #-}
-    put_ _ v = ReaderIORef $ do
-      ref <- ask @tag
-      liftIO $ writeIORef ref v
-    {-# INLINE put_ #-}
     state_ _ f = ReaderIORef $ do
       ref <- ask @tag
       liftIO $ atomicModifyIORef' ref (swap . f)
@@ -175,38 +123,11 @@ instance
         swap (a, b) = (b, a)
     {-# INLINE state_ #-}
 
--- | Derive a state monad from a reader over a mutable reference.
---
--- Mutable references are available in a 'Control.Monad.Primitive.PrimMonad'.
--- The corresponding 'Control.Monad.Primitive.PrimState' has to match the
--- 'Data.Mutable.MCState' of the reference. This constraint makes a stand-alone
--- deriving clause necessary.
---
--- Example:
---
--- > newtype MyState m a = MyState (ReaderT (IORef Int) m a)
--- >   deriving (Functor, Applicative, Monad)
--- > deriving via ReaderRef (MonadReader (ReaderT (IORef Int) m))
--- >   instance (PrimMonad m, PrimState m ~ PrimState IO)
--- >   => HasState "foo" Int (MyState m)
---
--- See 'ReaderIORef' for a specialized version over 'Data.IORef.IORef'.
-newtype ReaderRef m (a :: *) = ReaderRef (m a)
-  deriving (Functor, Applicative, Monad, MonadIO, PrimMonad)
-
 instance
   ( MutableRef ref, RefElement ref ~ s
   , HasReader tag ref m, PrimMonad m, PrimState m ~ MCState ref )
   => HasState tag s (ReaderRef m)
   where
-    get_ _ = ReaderRef $ do
-      ref <- ask @tag
-      readRef ref
-    {-# INLINE get_ #-}
-    put_ _ v = ReaderRef $ do
-      ref <- ask @tag
-      writeRef ref v
-    {-# INLINE put_ #-}
     state_ _ f = ReaderRef $ do
       ref <- ask @tag
       s <- readRef ref