+++ ChangeLog.md
@@ -1,5 +1,17 @@
 # Revision history for capability
 
+## 0.?.?.? -- ??
+
+* Rename HasStream to HasSink, for symmetry.
+  See [#75](https://github.com/tweag/capability/pull/75)
+
+* Introduce HasSource, a superclass of HasReader.
+  See [#75](https://github.com/tweag/capability/pull/75)
+
+* Make HasSource and HasSink superclasses of HasState.
+  See [#75](https://github.com/tweag/capability/pull/75)
+
+
 ## 0.2.0.0 -- 2019-03-22
 
 * Make HasStream a superclass of HasWriter.